Support direct links with authentication flow
Uses a different URI for direct links to preserve place location in WebProtege when authentication is required.  Also unifies the code for working around authentication when sessions expire with this functionality.  This addresses some comments in: https://github.com/who-icatx/icatx-project/issues/202 and unifies the approach with code that deals with https://github.com/who-icatx/icatx-project/issues/208.